rename namespace to mxnet and resolve naming clash
