+++ horovod/mxnet/cuda_util.cc
@@ -23,7 +23,7 @@
 #include "util.h"
 
 namespace horovod {
-namespace MX {
+namespace mxnet {
 
 with_device::with_device(int device) {
   if (device == CPU_DEVICE_ID) {
@@ -47,5 +47,5 @@ with_device::~with_device() {
 #endif
 }
 
-} // namespace MX
+} // namespace mxnet
 } // namespace horovod
