+++ horovod/mxnet/mpi_ops.h
@@ -21,11 +21,9 @@
 #include <mxnet/ndarray.h>
 
 namespace horovod {
-namespace MX {
+namespace mxnet {
 
 using namespace horovod::common;
-
-typedef mxnet::NDArray NDArray;
 
 extern "C" int horovod_mxnet_allreduce_async(
     NDArray* tensor, NDArray* output, int average, char* name);
@@ -36,7 +34,7 @@ extern "C" int horovod_mxnet_broadcast_async(
 extern "C" int horovod_mxnet_poll(int handle);
 extern "C" void horovod_mxnet_wait_and_clear(int handle);
 
-} // namespace MX
+} // namespace mxnet
 } // namespace horovod
 
 #endif // HOROVOD_MXNET_MPI_OPS_H
