+++ horovod/mxnet/ready_event.h
@@ -27,10 +27,10 @@
 #include "../common/common.h"
 
 namespace horovod {
-namespace MX {
+namespace mxnet {
 
 using namespace horovod::common;
-typedef mxnet::NDArray NDArray;
+typedef ::mxnet::NDArray NDArray;
 
 template <class T> class MXReadyEvent : public ReadyEvent {
 public:
@@ -42,7 +42,7 @@ private:
   NDArray* tensor_;
 };
 
-} // namespace MX
+} // namespace mxnet
 } // namespace horovod
 #endif
